refactor: rename AGP to SLIM
