+++ data-plane/core/controller/src/service.rs
@@ -28,7 +28,7 @@ use slim_config::grpc::client::ClientConfig;
 use slim_datapath::message_processing::MessageProcessor;
 use slim_datapath::messages::utils::SlimHeaderFlags;
 use slim_datapath::messages::{Agent, AgentType};
-use slim_datapath::pubsub::proto::pubsub::v1::Message as PubsubMessage;
+use slim_datapath::api::proto::pubsub::v1::Message as PubsubMessage;
 use slim_datapath::tables::SubscriptionTable;
 
 #[derive(Debug, Clone)]
