+++ data-plane/core/datapath/build.rs
@@ -12,7 +12,7 @@ fn main() {
     }
 
     tonic_build::configure()
-        .out_dir("src/pubsub/gen")
+        .out_dir("src/api/gen")
         .compile_protos(&["proto/v1/pubsub.proto"], &["proto/v1"])
         .unwrap();
 }
