+++ data-plane/core/datapath/src/message_processing.rs
@@ -21,11 +21,11 @@ use crate::connection::{Channel, Connection, Type as ConnectionType};
 use crate::errors::DataPathError;
 use crate::forwarder::Forwarder;
 use crate::messages::AgentType;
-use crate::pubsub::proto::pubsub::v1::message::MessageType::Publish as PublishType;
-use crate::pubsub::proto::pubsub::v1::message::MessageType::Subscribe as SubscribeType;
-use crate::pubsub::proto::pubsub::v1::message::MessageType::Unsubscribe as UnsubscribeType;
-use crate::pubsub::proto::pubsub::v1::pub_sub_service_client::PubSubServiceClient;
-use crate::pubsub::proto::pubsub::v1::{Message, pub_sub_service_server::PubSubService};
+use crate::api::proto::pubsub::v1::message::MessageType::Publish as PublishType;
+use crate::api::proto::pubsub::v1::message::MessageType::Subscribe as SubscribeType;
+use crate::api::proto::pubsub::v1::message::MessageType::Unsubscribe as UnsubscribeType;
+use crate::api::proto::pubsub::v1::pub_sub_service_client::PubSubServiceClient;
+use crate::api::proto::pubsub::v1::{Message, pub_sub_service_server::PubSubService};
 use crate::tables::connection_table::ConnectionTable;
 use crate::tables::subscription_table::SubscriptionTableImpl;
