+++ data-plane/core/service/src/producer_buffer.rs
@@ -3,7 +3,7 @@
 
 use std::collections::HashMap;
 
-use slim_datapath::{messages::AgentType, pubsub::proto::pubsub::v1::Message};
+use slim_datapath::{messages::AgentType, api::proto::pubsub::v1::Message};
 
 pub struct ProducerBuffer {
     capacity: usize,
@@ -92,8 +92,8 @@ impl ProducerBuffer {
 mod tests {
     use super::*;
     use slim_datapath::messages::encoder::{Agent, AgentType};
-    use slim_datapath::pubsub::proto::pubsub::v1::SessionHeaderType;
-    use slim_datapath::pubsub::{SessionHeader, SlimHeader};
+    use slim_datapath::api::proto::pubsub::v1::SessionHeaderType;
+    use slim_datapath::api::{SessionHeader, SlimHeader};
 
     #[test]
     fn test_producer_buffer() {
