+++ data-plane/core/service/src/session.rs
@@ -10,7 +10,7 @@ use crate::fire_and_forget::FireAndForgetConfiguration;
 use crate::request_response::RequestResponseConfiguration;
 use crate::streaming::StreamingConfiguration;
 use slim_datapath::messages::encoder::Agent;
-use slim_datapath::pubsub::proto::pubsub::v1::{Message, SessionHeaderType};
+use slim_datapath::api::proto::pubsub::v1::{Message, SessionHeaderType};
 
 /// Session ID
 pub type Id = u32;
